extracted common component barrier card
