+++ pages/review-barrier.tsx
@@ -4,127 +4,7 @@ import { useIntl } from "react-intl";
 import Button from "../components/Button";
 import Layout from "../components/Layout";
 import Page, { LeftSection, RightSection } from "../components/Page";
-import { PaperClipIcon } from "@heroicons/react/solid";
-
-const BarrierCard: React.FunctionComponent<{
-  title: string;
-  documentName: string;
-}> = ({ title, documentName }) => {
-  const intl = useIntl();
-  const bold = (msg: string): React.ReactNode => (
-    <span data-h2-font-weight="b(700)">{msg}</span>
-  );
-  return (
-    <section
-      data-h2-border="b(darkgray, left, solid, m)"
-      data-h2-margin="b(top, s)"
-      data-h2-shadow="b(m)"
-      data-h2-padding="b(all, m)"
-    >
-      <div
-        data-h2-display="b(flex)"
-        data-h2-justify-content="b(space-between)"
-        data-h2-margin="b(bottom, s)"
-      >
-        <p
-          data-h2-font-size="b(h5)"
-          data-h2-margin="b(all, none) b(right, m)"
-          data-h2-font-style="b(underline)"
-        >
-          {title}
-        </p>
-        <div>
-          <Link href="#">
-            <a data-h2-margin="b(right, s)">
-              {intl.formatMessage({
-                defaultMessage: "View",
-              })}
-            </a>
-          </Link>
-
-          <Link href="#">
-            <a data-h2-margin="b(right, s)">
-              {intl.formatMessage({
-                defaultMessage: "Edit",
-              })}
-            </a>
-          </Link>
-
-          <Link href="#">
-            <a>
-              {intl.formatMessage({
-                defaultMessage: "Delete",
-              })}
-            </a>
-          </Link>
-        </div>
-      </div>
-      <div
-        data-h2-display="b(flex)"
-        data-h2-justify-content="b(space-between)"
-        data-h2-align-items="b(center)"
-        data-h2-bg-color="b(lightblue)"
-      >
-        <p data-h2-margin="b(left, s) b(right, m)">
-          {intl.formatMessage({
-            defaultMessage:
-              "This solution is awaiting confirmation from your manager.",
-          })}
-        </p>
-        <p data-h2-margin="b(right, s)">
-          <Link href="#">
-            <a>
-              {intl.formatMessage({
-                defaultMessage: "Review",
-              })}
-            </a>
-          </Link>
-        </p>
-      </div>
-      <p data-h2-margin="b(bottom, none)">
-        {intl.formatMessage({
-          defaultMessage:
-            "This solution applies to me in following situations:",
-        })}
-      </p>
-      <ul data-h2-margin="b(bottom-left, m)" style={{ listStyleType: "disc" }}>
-        <li>
-          <p data-h2-margin="b(all, none)">
-            {intl.formatMessage({
-              defaultMessage: "Office Workplace",
-            })}
-          </p>
-        </li>
-        <li>
-          <p data-h2-margin="b(all, none)">
-            {intl.formatMessage({
-              defaultMessage: "Remote Workplace",
-            })}
-          </p>
-        </li>
-      </ul>
-      <div>
-        <p>
-          {intl.formatMessage(
-            {
-              defaultMessage:
-                "{title} is particularly helpful for me because...",
-            },
-            { title },
-          )}
-        </p>
-        <div data-h2-display="b(flex)">
-          <span data-h2-margin="b(right, xs)" data-h2-display="b(flex)">
-            <PaperClipIcon style={{ width: "1.2rem" }} />
-          </span>
-          <Link data-h2-margin="b(left, s)" href="#">
-            <a>{documentName}</a>
-          </Link>
-        </div>
-      </div>
-    </section>
-  );
-};
+import { BarrierCard } from "../components/formComponents/BarrierCard";
 
 const ReviewBarrier: React.FunctionComponent = () => {
   const intl = useIntl();
@@ -195,6 +75,7 @@ const ReviewBarrier: React.FunctionComponent = () => {
               documentName={intl.formatMessage({
                 defaultMessage: "my_hearing_assessment.pdf(3MB)",
               })}
+              actionLinks={["View", "Edit", "Delete"]}
             />
             <BarrierCard
               title={intl.formatMessage({
@@ -203,6 +84,7 @@ const ReviewBarrier: React.FunctionComponent = () => {
               documentName={intl.formatMessage({
                 defaultMessage: "location_assessment.pdf(3MB)",
               })}
+              actionLinks={["View", "Edit", "Delete"]}
             />
             <BarrierCard
               title={intl.formatMessage({
@@ -212,6 +94,7 @@ const ReviewBarrier: React.FunctionComponent = () => {
               documentName={intl.formatMessage({
                 defaultMessage: "software_assessment.pdf(3MB)",
               })}
+              actionLinks={["View", "Edit", "Delete"]}
             />
             <h4 data-h2-font-size="b(h4)" data-h2-margin="b(bottom, none)">
               {intl.formatMessage(
